// app/types.ts
export interface ReqOutput {
  ip?: string;
  continent?: string;
  continentCode?: string;
  country?: string;
  countryCode?: string;
  countryEmoji?: string;
  region?: string;
  regionCode?: string;
  city?: string;
  connectionAsn?: number;
  connectionOrg?: string;
  connectionIsp?: string;
  connectionDomain?: string;
}